All ID functions can now be done by reply.
I should have done this in the first place

+++ promote.php
@@ -17,7 +17,7 @@
     You should have received a copy of the GNU General Public License
     along with BruhhBot. If not, see <http://www.gnu.org/licenses/>.
  */
-function promoteme($update, $MadelineProto, $msg)
+function promoteme($update, $MadelineProto, $msg = "")
 {
     if (is_supergroup($update, $MadelineProto)) {
         $msg_id = $update['update']['message']['id'];
@@ -33,16 +33,37 @@ function promoteme($update, $MadelineProto, $msg)
             );
         if (is_moderated($ch_id)) {
             if (from_admin($update, $MadelineProto, $mods, true)) {
-                $id = catch_id($update, $MadelineProto, $msg);
-                if ($id[0]) {
-                    $userid = $id[1];
-                    $username = $id[2];
-                    $mention = html_mention($username, $userid);
-                    check_json_array('promoted.json', $ch_id);
-                    $file = file_get_contents("promoted.json");
-                    $promoted = json_decode($file, true);
-                    if (array_key_exists($ch_id, $promoted)) {
-                        if (!in_array($userid, $promoted[$ch_id])) {
+                if (!empty($msg) or array_key_exists('reply_to_msg_id', $update['update']['message'])) {
+                    $id = catch_id($update, $MadelineProto, $msg);
+                    if ($id[0]) {
+                        $userid = $id[1];
+                        $username = $id[2];
+                        $mention = html_mention($username, $userid);
+                        check_json_array('promoted.json', $ch_id);
+                        $file = file_get_contents("promoted.json");
+                        $promoted = json_decode($file, true);
+                        if (array_key_exists($ch_id, $promoted)) {
+                            if (!in_array($userid, $promoted[$ch_id])) {
+                                array_push($promoted[$ch_id], $userid);
+                                file_put_contents('promoted.json', json_encode($promoted));
+                                $str = $MadelineProto->responses['promoteme']['success'];
+                                $repl = array(
+                                    "mention" => $mention,
+                                    "title" => $title
+                                );
+                                $message = $MadelineProto->engine->render($str, $repl);
+                                $default['message'] = $message;
+                            } else {
+                                $str = $MadelineProto->responses['promoteme']['already'];
+                                $repl = array(
+                                    "mention" => $mention,
+                                    "title" => $title
+                                );
+                                $message = $MadelineProto->engine->render($str, $repl);
+                                $default['message'] = $message;
+                            }
+                        } else {
+                            $promoted[$ch_id] = [];
                             array_push($promoted[$ch_id], $userid);
                             file_put_contents('promoted.json', json_encode($promoted));
                             $str = $MadelineProto->responses['promoteme']['success'];
@@ -52,34 +73,15 @@ function promoteme($update, $MadelineProto, $msg)
                             );
                             $message = $MadelineProto->engine->render($str, $repl);
                             $default['message'] = $message;
-                        } else {
-                            $str = $MadelineProto->responses['promoteme']['already'];
-                            $repl = array(
-                                "mention" => $mention,
-                                "title" => $title
-                            );
-                            $message = $MadelineProto->engine->render($str, $repl);
-                            $default['message'] = $message;
                         }
                     } else {
-                        $promoted[$ch_id] = [];
-                        array_push($promoted[$ch_id], $userid);
-                        file_put_contents('promoted.json', json_encode($promoted));
-                        $str = $MadelineProto->responses['promoteme']['success'];
+                        $str = $MadelineProto->responses['promoteme']['idk'];
                         $repl = array(
-                            "mention" => $mention,
-                            "title" => $title
+                            "msg" => $msg
                         );
                         $message = $MadelineProto->engine->render($str, $repl);
                         $default['message'] = $message;
                     }
-                } else {
-                    $str = $MadelineProto->responses['promoteme']['idk'];
-                    $repl = array(
-                        "msg" => $msg
-                    );
-                    $message = $MadelineProto->engine->render($str, $repl);
-                    $default['message'] = $message;
                 }
             }
         }
@@ -94,7 +96,7 @@ function promoteme($update, $MadelineProto, $msg)
     }
 }
 
-function demoteme($update, $MadelineProto, $msg)
+function demoteme($update, $MadelineProto, $msg = "")
 {
     if (is_supergroup($update, $MadelineProto)) {
         $msg_id = $update['update']['message']['id'];
@@ -109,33 +111,43 @@ function demoteme($update, $MadelineProto, $msg)
             'parse_mode' => 'html'
             );
         if (from_admin($update, $MadelineProto, $mods, true)) {
-            $id = catch_id($update, $MadelineProto, $msg);
-            if ($id[0]) {
-                $userid = $id[1];
-                $username = $id[2];
-                $mention = html_mention($username, $userid);
-                check_json_array('promoted.json', $ch_id);
-                $file = file_get_contents("promoted.json");
-                $promoted = json_decode($file, true);
-                if (array_key_exists($ch_id, $promoted)) {
-                    if (in_array($userid, $promoted[$ch_id])) {
-                        if (($key = array_search(
-                            $userid,
-                            $promoted[$ch_id]
-                        )) !== false
-                        ) {
-                            unset($promoted[$ch_id][$key]);
+            if (!empty($msg) or array_key_exists('reply_to_msg_id', $update['update']['message'])) {
+                $id = catch_id($update, $MadelineProto, $msg);
+                if ($id[0]) {
+                    $userid = $id[1];
+                    $username = $id[2];
+                    $mention = html_mention($username, $userid);
+                    check_json_array('promoted.json', $ch_id);
+                    $file = file_get_contents("promoted.json");
+                    $promoted = json_decode($file, true);
+                    if (array_key_exists($ch_id, $promoted)) {
+                        if (in_array($userid, $promoted[$ch_id])) {
+                            if (($key = array_search(
+                                $userid,
+                                $promoted[$ch_id]
+                            )) !== false
+                            ) {
+                                unset($promoted[$ch_id][$key]);
+                            }
+                            file_put_contents('promoted.json', json_encode($promoted));
+                            $str = $MadelineProto->responses['demoteme']['success'];
+                            $repl = array(
+                                "mention" => $mention,
+                                "title" => $title
+                            );
+                            $message = $MadelineProto->engine->render($str, $repl);
+                            $default['message'] = $message;
+                        } else {
+                            $str = $MadelineProto->responses['demoteme']['fail'];
+                            $repl = array(
+                                "mention" => $mention,
+                                "title" => $title
+                            );
+                            $message = $MadelineProto->engine->render($str, $repl);
+                            $default['message'] = $message;
                         }
-                        file_put_contents('promoted.json', json_encode($promoted));
-                        $str = $MadelineProto->responses['demoteme']['success'];
-                        $repl = array(
-                            "mention" => $mention,
-                            "title" => $title
-                        );
-                        $message = $MadelineProto->engine->render($str, $repl);
-                        $default['message'] = $message;
                     } else {
-                        $str = $MadelineProto->responses['demoteme']['fail'];
+                        $str = $MadelineProto->responses['demoteme']['success'];
                         $repl = array(
                             "mention" => $mention,
                             "title" => $title
@@ -144,21 +156,13 @@ function demoteme($update, $MadelineProto, $msg)
                         $default['message'] = $message;
                     }
                 } else {
-                    $str = $MadelineProto->responses['demoteme']['success'];
+                    $str = $MadelineProto->responses['demoteme']['idk'];
                     $repl = array(
-                        "mention" => $mention,
-                        "title" => $title
+                        "msg" => $msg
                     );
                     $message = $MadelineProto->engine->render($str, $repl);
                     $default['message'] = $message;
                 }
-            } else {
-                $str = $MadelineProto->responses['demoteme']['idk'];
-                $repl = array(
-                    "msg" => $msg
-                );
-                $message = $MadelineProto->engine->render($str, $repl);
-                $default['message'] = $message;
             }
         }
         if (isset($default['message'])) {
